welfare-crawler/requirements.txt
requests==2.32.3
mysql-connector-python==8.4.0
lxml==5.2.1     